% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpROC-package.R
\docType{package}
\name{fpROC-package}
\alias{fpROC}
\alias{fpROC-package}
\title{fpROC: Fast partial ROC test}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The package features optimized C++ code for computing partial ROC. Parallelization using OpenMP (OMP) was achieved with assistance from the DeepSeek AI Assistant (https://deepseek.com).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://luismurao.github.io/fpROC/}
  \item Report bugs at \url{https://github.com/luismurao/fpROC/issues}
}

}
\author{
\strong{Maintainer}: Luis Osorio-Olvera \email{luismurao@gmail.com} (\href{https://orcid.org/0000-0003-0701-5398}{ORCID})

Authors:
\itemize{
  \item Marlon E. Cobos \email{manubio13@gmail.com} (\href{https://orcid.org/0000-0002-2611-1767}{ORCID})
  \item Rusby G. Contreras-Díaz \email{rusby.contreras.diaz@gmail.com} (\href{https://orcid.org/0000-0002-0569-8984}{ORCID})
}

Other contributors:
\itemize{
  \item Weverton Trindade \email{wevertonf1993@gmail.com} (\href{https://orcid.org/0000-0003-2045-4555}{ORCID}) [contributor]
  \item Luis F. Arias-Giraldo \email{lfarias.giraldo@gmail.com} (\href{https://orcid.org/0000-0003-4861-8064}{ORCID}) [contributor]
}

}
\keyword{internal}
